homework 7 #7
