+++ main.js
@@ -1,98 +1,31 @@
-const $arenas = document.querySelector('.arenas')
-const $fightButton = document.querySelector('.button')
-const $formFight = document.querySelector('.control')
-const $chat = document.querySelector('.chat')
+import {player1, player2} from "./player.js";
+import {generateLogs} from "./logs.js";
+import {createElement} from "./createElement.js";
+import {enemyAttack, playerAttack} from "./attack.js";
+import {finishFight} from "./finishFight.js";
 
-const HIT = {
-    head: 30,
-    body: 25,
-    foot: 20,
-}
-const ATTACK = ['head', 'body', 'foot'];
+/* variables */
+export const $arenas = document.querySelector('.arenas'),
+    $fightButton = document.querySelector('.button'),
+    $formFight = document.querySelector('.control'),
+    $chat = document.querySelector('.chat'),
 
-const logs = {
-    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-    end: [
-        'Результат удара [playerWins]: [playerLose] - труп',
-        '[playerLose] погиб от удара бойца [playerWins]',
-        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-    ],
-    hit: [
-        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-    ],
-    defence: [
-        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-    ],
-    draw: 'Ничья - это тоже победа!'
-};
+    ATTACK = ['head', 'body', 'foot'],
 
-let player1 = {
-    player: 1,
-    name: 'Scorpion',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
-    weapon: ['Chain', 'Kunai'],
-    attack: function () {
-        console.log(name + 'Fight...')
-    },
-    elHP,
-    changeHP,
-    renderHP,
-}
+    HIT = {
+        head: 30,
+        body: 25,
+        foot: 20,
+    };
+/* end of variables */
 
-let player2 = {
-    player: 2,
-    name: 'Sub-Zero',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
-    weapon: ['Ice', 'Shuriken'],
-    attack: function () {
-        console.log(name + 'Fight...')
-    },
-    elHP,
-    changeHP,
-    renderHP,
-}
-
-function createElement(tag, className) {
-    const $tag = document.createElement(tag)
-    if (className) {
-        $tag.classList.add(className)
-    }
-    return $tag
-}
-
-function createPlayer(playerObj) {
-    const $player = createElement('div', 'player' + playerObj.player)
-    const $progressbar = createElement('div', 'progressbar')
-    const $life = createElement('div', 'life')
-    const $name = createElement('div', 'name')
-    const $character = createElement('div', 'character')
-    const $img = createElement('img')
+const createPlayer = (playerObj) => {
+    const $player = createElement('div', 'player' + playerObj.player),
+        $progressbar = createElement('div', 'progressbar'),
+        $life = createElement('div', 'life'),
+        $name = createElement('div', 'name'),
+        $character = createElement('div', 'character'),
+        $img = createElement('img')
 
     $life.style.width = playerObj.hp + '%'
     $name.innerText = playerObj.name
@@ -106,149 +39,10 @@ function createPlayer(playerObj) {
     return $player
 }
 
-function playerWin(name) {
-    const $winTitle = createElement('div', 'winTitle')
-    if (name) {
-        $winTitle.innerText = name + ' wins'
-    } else {
-        $winTitle.innerText = 'draw '
-    }
-    return $winTitle
-}
-
-function getRandomNum(num) {
-    return Math.floor(Math.random() * num)
-}
-
-function changeHP(value) {
-    this.hp -= value
-    if (this.hp <= 0) {
-        this.hp = 0
-    }
-    this.renderHP()
-}
-
-function elHP() {
-    return document.querySelector(`.player${this.player} .life`)
-}
-
-function renderHP() {
-    this.elHP().style.width = `${this.hp}%`
-}
-
-function createReloadButton() {
-    const $reloadButton = createElement('div', 'reloadWrap')
-    const $button = createElement('button', 'button')
-    $button.innerText = 'Restart'
-    $reloadButton.appendChild($button)
-    $arenas.appendChild($reloadButton)
-
-    $reloadButton.addEventListener('click', function () {
-        window.location.reload()
-    }, {once: true})
-}
-
-// let playerArr = [player1, player2]
-
-/*function randomPlayer(arr) {
-    const index = Math.floor(Math.random() * 2)
-    console.log(arr[index])
-    return arr[index]
-}*/
-
 $arenas.appendChild(createPlayer(player1))
 $arenas.appendChild(createPlayer(player2))
 
-function enemyAttack() {
-    const hit = ATTACK[getRandomNum(3)]
-    const defence = ATTACK[getRandomNum(3)]
-
-    return {
-        value: getRandomNum(HIT[hit]),
-        hit,
-        defence,
-    }
-}
-
-function playerAttack() {
-    const attack = {};
-
-    for (let item of $formFight) {
-        if (item.checked && item.name === 'hit') {
-            attack.value = getRandomNum(HIT[item.value])
-            attack.hit = item.value
-        }
-        if (item.checked && item.name === 'defence') {
-            attack.defence = item.value
-        }
-        item.checked = false
-    }
-    return attack;
-}
-
-function getData() {
-    const date = new Date()
-    return `${date.getHours()}:${date.getMinutes()}`
-}
-
-function generateLogs(type, player1, player2) {
-    switch (type) {
-        case 'start': {
-            const text = logs[type].replace('[time]', getData()).replace('[player1]', player1.name).replace('[player2]', player2.name)
-            const el = `<p>${text}</p>`
-            $chat.insertAdjacentHTML('afterbegin', el)
-            break;
-        }
-        case 'hit': {
-            const text = `${getData()} ${logs[type][getRandomNum(logs.hit.length)].replace('[playerDefence]', player1.name).replace('[playerKick]', player2.name)} / ${player1.name} - ${player1.hp}HP`
-            const el = `<p>${text}</p>`
-            $chat.insertAdjacentHTML('afterbegin', el)
-            break;
-        }
-        case 'defence': {
-            const text = `${getData()} ${logs[type][getRandomNum(logs.defence.length)].replace('[playerKick]', player1.name).replace('[playerDefence]', player2.name)}`
-            const el = `<p>${text}</p>`
-            $chat.insertAdjacentHTML('afterbegin', el)
-            break;
-        }
-        case 'end': {
-            const text = `${getData()} ${logs[type][getRandomNum(logs.end.length)].replace('[playerWins]', player1.name).replace('[playerLose]', player2.name)}`
-            const el = `<p>${text}</p>`
-            $chat.insertAdjacentHTML('afterbegin', el)
-            break;
-        }
-        case 'draw': {
-            const text = `${getData()} ${logs[type]}`
-            const el = `<p>${text}</p>`
-            $chat.insertAdjacentHTML('afterbegin', el)
-            break;
-        }
-    }
-}
-
-function finishFight() {
-
-    if (player1.hp === 0 || player2.hp === 0) {
-        $fightButton.disabled = true
-        createReloadButton()
-    }
-
-    if (player1.hp === 0 && player1.hp < player2.hp) {
-        $arenas.appendChild(playerWin(player2.name))
-        generateLogs('end', player2, player1)
-    }
-    if (player2.hp === 0 && player2.hp < player1.hp) {
-        $arenas.appendChild(playerWin(player1.name))
-        generateLogs('end', player1, player2)
-    }
-    if (player1.hp === 0 && player2.hp === 0) {
-        $arenas.appendChild(playerWin())
-        generateLogs('draw', player1, player2)
-    }
-
-}
-
-$formFight.addEventListener('submit', function (e) {
+$formFight.addEventListener('submit', (e) => {
     e.preventDefault()
     const enemy = enemyAttack()
     const player = playerAttack()
@@ -256,19 +50,15 @@ $formFight.addEventListener('submit', function (e) {
     if (player.hit !== enemy.defence) {
         player2.changeHP(player.value)
         generateLogs('hit', player2, player1)
-        console.log('HIT')
     } else {
         generateLogs('defence', player1, player2)
-        console.log('DEFENCE')
     }
 
     if (enemy.hit !== player.defence) {
         player1.changeHP(enemy.value)
         generateLogs('hit', player1, player2)
-        console.log('HIT')
     } else {
         generateLogs('defence', player2, player1)
-        console.log('DEFENCE')
     }
 
     finishFight()
@@ -277,4 +67,5 @@ $formFight.addEventListener('submit', function (e) {
 function init() {
     generateLogs('start', player1, player2)
 }
+
 init()
