+++ utils.js
@@ -1,0 +1,9 @@
+export const getRandomNum = (num) => {
+    return Math.floor(Math.random() * num)
+}
+
+export const getData = () => {
+    const date = new Date()
+    return `${date.getHours()}:${date.getMinutes()}`
+}
+
